+++ src/code/z_collision_check.c
@@ -16,7 +16,7 @@
 #include "z_lib.h"
 
 #pragma increment_block_number "gc-eu:128 gc-eu-mq:128 gc-jp:128 gc-jp-ce:128 gc-jp-mq:128 gc-us:128 gc-us-mq:128" \
-                               "ntsc-1.0:88 ntsc-1.1:88 ntsc-1.2:88 pal-1.0:88 pal-1.1:88"
+                               "ntsc-1.0:96 ntsc-1.1:88 ntsc-1.2:88 pal-1.0:88 pal-1.1:88"
 
 typedef s32 (*ColChkResetFunc)(PlayState*, Collider*);
 typedef void (*ColChkApplyFunc)(PlayState*, CollisionCheckContext*, Collider*);
